Refactor source registry
