+++ lilac/data/dataset_config_test.py
@@ -15,7 +15,7 @@ from ..config import (
 )
 from ..schema import Field, Item, RichData, field, lilac_embedding
 from ..signal import TextEmbeddingSignal, TextSignal, clear_signal_registry, register_signal
-from ..sources.source_registry import clear_source_registry, register_source
+from ..source import clear_source_registry, register_source
 from .dataset_test_utils import TestDataMaker, TestSource
