+++ lilac/load_dataset_test.py
@@ -15,8 +15,7 @@ from .data.dataset_utils import get_parquet_filename
 from .load_dataset import process_source
 from .project import read_project_config
 from .schema import PARQUET_FILENAME_PREFIX, ROWID, Item, schema
-from .source import Source, SourceManifest, SourceSchema
-from .sources.source_registry import clear_source_registry, register_source
+from .source import Source, SourceManifest, SourceSchema, clear_source_registry, register_source
 from .test_utils import fake_uuid, read_items
 from .utils import DATASETS_DIR_NAME
