+++ lilac/source_test.py
@@ -4,9 +4,8 @@ from typing import ClassVar, Iterable, cast
 import pytest
 from typing_extensions import override
 
-from ..schema import Item
-from ..source import Source, SourceSchema
-from .source_registry import clear_source_registry, get_source_cls, register_source, resolve_source
+from .schema import Item
+from .source import Source, SourceSchema, clear_source_registry, get_source_cls, register_source, resolve_source
 
 
 class TestSource(Source):
